Fix review: - Move BlobImpl implementation to Blob - Remove BlobImpl
